Moved command_line module to main.rs
