--- src/command_line_parser/mod.rs
@@ -1,1 +1,0 @@
-pub mod parser;
