--- src/command_line_parser/parser.rs
@@ -1,17 +1,0 @@
-use clap::{arg, ArgGroup, ArgMatches, Command};
-
-pub fn receive_command_line_arguments() -> Result<ArgMatches, String> {
-    let args = Command::new("LC3 Virtual Machine")
-        .arg(arg!(-i --interactive "interactive console").required(false))
-        .arg(arg!(-f --file <FILE> "file to execute").required(false))
-        .arg(arg!(-d --debug <FILE> "debug file").required(false))
-        .group(
-            ArgGroup::new("run program")
-                .args(["interactive", "file", "debug"])
-                .required(false),
-        )
-        .after_help("Don't use -i, -f or -d at the same time")
-        .get_matches();
-
-    Ok(args)
-}
