+++ src/lib.rs
@@ -1,4 +1,3 @@
-pub mod command_line_parser;
 pub mod constants;
 pub mod virtual_machine;
 pub mod virtual_machine_start;
